bump: Replace unmaintained dependency serde_yml with serde_yaml_ng
We've been using serde_yml as a replacement for the abandoned crate
serde_yaml, but now it's been abandoned as well. The corresponding
RustSec advisory suggests two alternatives, serde_yaml_ng is one of
them. It's a fork of serde_yaml, too, and seems to work well as a
drop-in successor.

There are a couple of crates where we don't actually use the serde_yml
dependency included from their Cargo.toml: just remove the Cargo.toml
entries for those. I'm not sure we still need the Serialize/Deserialize
implementations for Prefix in the lpm crate, for which we have the one
test that actually requires serde_yaml_ng in the code base, but I left
it for now.

Link: https://rustsec.org/advisories/RUSTSEC-2025-0067
Link: https://rustsec.org/advisories/RUSTSEC-2025-0068
Signed-off-by: Quentin Monnet <qmo@qmon.net>