refactor(nat): Make stateful NAT's translate_packet IP version-agnostic
It's painful to have two versions translate_packet_v4() and
translate_packet_v6(), nearly identical, to update all the time. By
extending trait NatIpWithBitmap a little, we can merge the two into a
single generic translate_packet() function.

Ideally, the allocate() method should be added to the more generic NatIp
trait, but this requires a new trait for the returned type, because type
AllocatedIpPort is tied to the NatIpWithBitmap trait. This will be for
another time.

Signed-off-by: Quentin Monnet <qmo@qmon.net>