refactor(nat): Move NatState type to stateful module, remove sessions.rs
Submodule "sessions" was mostly emptied when we switched to the new flow
table implementation, then removed the NatTuple type. There's only the
NatState type left; move it to the main "stateful" module, in a
simplified version. We'll change it in a follow-up commit to directly
embed the objects from the allocator into the flow table.

Signed-off-by: Quentin Monnet <qmo@qmon.net>