feat(nat): Store AllocatedIpPort object in flow table for stateful NAT
Rather than storing a NatTranslationData (previously known as NatState)
object in the flow table, introduce a new type that embeds to
AllocatedIpPort objects, and make it the value for entries in the flow
table.

This is important, because it ensures we the AllocatedIpPort objects
alive in the flow table; and with them, the back references to the
various objects created by the stateful NAT allocator, such that these
objects are only garbage-collected (dropped) when the entry is removed
from the table and dropped.

Signed-off-by: Quentin Monnet <qmo@qmon.net>