refactor(nat): Move L4 ports setters to net crate
The match on the transport layer protocol to set the source and
destination ports for TCP or UDP appropriately has little interest in
the stateful NAT code. To make the module clearer, move these to the net
crate, by introducing two new methods try_set_source() and
try_set_destination() for the Transport type.

Signed-off-by: Quentin Monnet <qmo@qmon.net>