feat(tracectl): keep a collection of tags
Tags represent collection of targets. Add an additional map for
tags as it eases certain operations. The map only keeps the names
of the target members and not their config (log-level). This is so
because if we'd keep the config, the config would need to be shared
by multiple tags in some Arc<Mutex> as a target may belong to more
than one tag. Otherwise, multiple objects for the same target with
distinct loglevels would exist. I'm not going for such an approach
now as it would become unnecessarily complex.

Also, add an auxiliary type TargetCfgDbByTag and implement Display
for it to be able to show targets grouped by tags.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>