feat(dataplane): add cmd line args for tracing
Add command line args related to tracing.
This is mostly for developing convenience: prior commits declared
multiple targets with logs with log-level < debug (on purpose to
reduce churn). Since an API to set the log-levels is not yet in
place, we need an easy way to enable logs for development.

This patch adds cmd line args to show the configurable targets
and their tags and to apply some tracing configuration from
a string composed of comma-separated "tag=level" substrings.

 --show-tracing-tags
     Show the available tracing tags and stop the dataplane
 --show-tracing-targets
     Show configurable tracing targets and stop the dataplane
 --tracing <tracing configuration>
     Tracing config string as comma-separated sequence of tag=level,
     with level one in [off,error,warn,info,debug,trace]

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>