feat(tracectl): set tracing config for N tags
Add a method to set the loglevel for a collection of tags at once.
If the provided tags are not disjoint (in terms of targets), the
common targets will get the highest loglevel among the tags it
belongs to. This is the agreed behavior in the API and is
justified by the fact that, in case of ambiguity, it is better
to have extra logs than possibly miss important logs.

This patch also defines an error type and renames the level field
of the tracing control db to "default" as it signifies the
default log-level.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>