refactor(net): Refactor parser calling for L3 -> L4 parsing
The code for calling parsers for IP, IP extensions, IP authentication is
pretty much the same, repeated many times accross different submodules.
Let's refactor it, to make future changes easier. We introduce a new
ParseHeader trait that we implement (with default methods) for the
different Ip*Next types, and call its methods everywhere necessary.

We don't really get a shorter line count at the end of the day, because
we need to implement the From trait for all possible variants for the
Ip*Next enums each time, but it remains clearer than the chunk we avoid
duplicating.

Signed-off-by: Quentin Monnet <qmo@qmon.net>