feat(tracectl): abstract policy on overlap
When configuring the tracing, if multiple tags overlap, it has to
be decided what loglevel is set for the targets in the
intersection. Currently, the policy is to set the highest of the
log-levels. However, we may want to change that policy since it is
not always convenient. For instance, in a configuration like
  pipeline => error
  nat => debug
this would set the loglevel of all members in the pipeline to
error, except for nat (also in pipeline) to debug possibly
matching users' intent. However, in a configuration like
  pipeline => debug,
  nat => off
meant to debug the pipeline except for nat, this would not
produce the desired result as the log-level for nat would remain
at debug.

This patch abstracts that decision by defining a trait so that
other policies can be implemented.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>