refactor(net): Use a macro to implement From trait for enum variants
In the net parser, we have a number of enum with members taking
arguments of a specific type, and for which we implement the From trait
from that type into the enum type. This is convenient for converting
parsed headers into the enum, but this leads to verbose implementations
of the trait for all enum variants.

This is easy enough to make it less verbose with a dedicated macro. We
can also use this macro for struct Header in headers.rs.

This is in prevision for implementing the From trait on even more enum
objects.

Signed-off-by: Quentin Monnet <qmo@qmon.net>