docs: Add comments for enum TruncatedTcp and TruncatedUdp variants
These comments are not exactly essential. At some point I thought I'd
have to make the variants publicly accessible, and the compiler asked
for the associated documentation; now I've written it, I figured I might
just as well commit it. The comments for FullHeader() does add a
precision that the header being full does not imply the payload is.

Signed-off-by: Quentin Monnet <qmo@qmon.net>