refactor(net): Rename ICMPv4-related traits to make version explicit
At the time ICMP was created, there was no IPv6 and no ICMPv6, so that
it's often called just ICMP when ICMPv4 might be more useful to lift the
confusion.

Given that we use ICMPv4, ICMPv6, and sometimes we want to deal with
ICMP in a generic fashion regardless of the version number, rename trait
TryIcmp and related traits and methods as TryIcmp4 (and accordingly) to
make ICMP version explicit and avoid ambiguity.

Signed-off-by: Quentin Monnet <qmo@qmon.net>