refactor: Make packet's headers access consistent
We've had two ways to access to a packet's immutable Headers structure:
via Packet's .get_headers(), and via .headers() from the TryHeaders
trait. To make things consistent, just use the .headers() method from
the TryHeaders trait everywhere, then delete the .get_headers() method.

... Oh wait, not so fast! Because it returns an AbstractHeaders
implementation, .headers() does not provide an object that implements
the Display trait. So let's keep .get_headers() so that we can call it
to implement Display for a Packet, but let's reduce its scope to the
crate, only.

Signed-off-by: Quentin Monnet <qmo@qmon.net>