refactor(net): Un-nest update_checksums() content
Rework function update_checksums() to make it (arguably) easier to
follow, and (certainly) easier to extend. Return early when we can, and
process each layer in successive steps, rather than having nested
destructuring. Also access VXLAN headers (if any) as immutable.

This is in preparation for extending the function with checksum updates
for IP packets embedded in ICMP.

There should be no functional change.

Signed-off-by: Quentin Monnet <qmo@qmon.net>