refactor(net): Remove Tcp's checksum/set_checksum, use Checksum trait
There are two conflicting checksum() and set_checksum() methods for
struct Tcp: those implemented directly for the struct, and those
available as part of the Checksum trait implementation. Even though
these implementations are currently identical, this may lead to
confusion; and in the future, we plan to change the methods signature
for the trait. Let's remove the "duplicate" methods implemented for
struct Tcp, and call those provided for the Checksum trait instead.

There should be no functional change.

Signed-off-by: Quentin Monnet <qmo@qmon.net>