feat(net): Implement AbstractEmbeddedHeaders & Co. for EmbeddedHeaders
For parsing inner IP packets embedded in ICMP Error messages, we have
struct EmbeddedHeaders, and it works well. Accessing the parsed headers
from that struct, though, requires a few more adjustments. We could make
all fields of the struct public and access them directly, but that
wouldn't be a clean solution. The alternative is to reproduce what's in
place for struct Headers, with its AbstractHeaders trait (and mutable
variant), and all the intermediary Try... traits (and mutable variants).

This is what we implement in this commit: We introduce and implement
AbstractEmbeddedHeaders (and mutable variants), and all the litany of
associated Try... traits (and, you guessed it, their mutable variants).

Worth of note are TryEmbeddedHeaders and TryEmbeddedHeadersMut, used for
bindings, but not implemented yet, and that provide access to the
EmbeddedHeaders of the object implementing them. These traits will be
implemented for struct Packet in a follow-up commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>