feat(net): Implement TryEmbeddedHeaders (and ...Mut) for struct Packet
In the previous commit we introduced traits TryEmbeddedHeaders and
TryEmbeddedHeadersMut to return an "impl AbstractEmbeddedHeaders" (or
its mutable variant) - in practice, to give access to the struct
EmbeddedHeaders from the calling object. Now, we implement them for
struct Packet, so that we can call packet.embedded_headers() and
packet.embedded_headers_mut().

Signed-off-by: Quentin Monnet <qmo@qmon.net>