feat(net): Implement checksum handling for EmbeddedTransport
Implement the Checksum trait for structs TruncatedTcp and TruncatedUdp,
then use the resulting methods to implement checksum() and
update_checksum() for struct EmbeddedTransport. This will be useful in
follow-up work to update checksums after translating ports in NAT.

We don't implement a method to validate the checksum for this header,
because we don't have a use case for it. From REQ-3, c) of RFC 5508:

    The NAT device SHOULD NOT validate the transport checksum of the
    embedded packet within an ICMP Error message, even when it is
    possible to do so;

Note that if we were to implement this checksum validation function in
the future, we should be very careful when passing the payload for
computing the checksum, as the buffer may contain a truncated payload,
or padding and ICMP Extension Structures at the end which should be left
out for computing the payload. TruncatedUdp and TruncatedTcp are headers
and have no knowledge of the payload and its supposed length, so their
methods just work with the payload provided by the caller, no questions
asked.

Link: https://datatracker.ietf.org/doc/html/rfc5508
Signed-off-by: Quentin Monnet <qmo@qmon.net>