feat(left-right-tlcache): avoid duplicate rhandles
The thread-local cache stores ReadHandle<T>'s based on some key,
where the key may match the identity or be an alias. Currently,
we store a distinct readhandle per key, regardless of whether
the readhandle allows reading from the same object. Exploit the
fact that we learn the identity of a T when querying the provider
to let cached entries for distinct keys but the same identity share
a single ReadHandle<T>. This should expedite internal operations
in case there would be many handles. In the sample use case, we'll
cache the same ReadHandle<Fib> for the two keys it may be queried
(vrfId or Vni).

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>